Authorizations

+++ api/app/Http/Controllers/CustomerController.php
@@ -19,10 +19,14 @@ class CustomerController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('can:viewAny')->only('viewAny');
-        $this->middleware('can:create')->only('create');
-        $this->middleware('can:update')->only('update');
-        $this->middleware('can:delete')->only('delete');
+        //$this->middleware('can:viewAny')->only('viewAny');
+        //$this->middleware('can:create')->only('create');
+        //$this->middleware('can:update')->only('update');
+        //$this->middleware('can:delete')->only('delete');
+        //$this->middleware('can:get_number_customers_created_this_month')->only('get_number_customers_created_this_month');
+        //$this->middleware('can:show_by_user')->only('show_by_user');
+
+
         /*
         $this->middleware('auth.manager', ['except' => [
             'store',
@@ -56,11 +60,14 @@ class CustomerController extends Controller
     }
     public function index()
     {
+        if (Auth()->guard('api')->user()->type != "EM") { abort(403); }
         return CustomerResource::collection(Customer::all());
     }
 
     public function store(StoreCustomerRequest $customer_request, StoreUserRequest $user_request)
     {
+        if (Auth()->guard('api')->user()->type != "C") { abort(403); }
+
         /* --- DB Transaction -> Create User + Customer --- */
         $new_customer = DB::transaction(function () use ($customer_request, $user_request) : Customer {
             // -> Creates User
@@ -89,11 +96,15 @@ class CustomerController extends Controller
 
     public function show(Customer $customer)
     {
+        if ((!Auth()->guard('api')->user()->type == "EM") && (Auth()->guard('api')->user()->type !="EM" && Auth()->guard('api')->user()->id!=$customer->id) ) { abort(403); }
+
         return new CustomerResource($customer);
     }
 
     public function update(StoreCustomerRequest $customer_request, UpdateUserRequest $user_request, Customer $customer)
     {
+        if (Auth()->guard('api')->user()->type != "C") { abort(403); }
+
         $updated_customer = DB::Transaction(function () use ($customer_request, $user_request, $customer) : Customer {
             // -> Updates User
             $updated_user = (new UserController)->update($user_request, $customer->user);
@@ -146,6 +157,8 @@ class CustomerController extends Controller
 
     public function destroy($id) // -> Boolean Return
     {
+        if (Auth()->guard('api')->user()->type != "EM") { abort(403); }
+
         return DB::transaction(function () use ($id) {
             $customer = Customer::where(['id' => $id], ['deleted_at' => null])->firstOrFail();
             $customer->user->delete();
@@ -171,5 +184,21 @@ class CustomerController extends Controller
     }
 
     public function show_by_user(User $user){ return new CustomerResource($user->customer); }
-
+/*
+    protected function resourceAbilityMap()
+    {
+        return array_merge(parent::resourceAbilityMap(), [
+            'get_number_customers_created_this_month' => 'get_number_customer_created_this_month',
+            'show_by_user' => 'show_by_user'
+    
+        ]);
+    }
+    protected function resourceMethodsWithoutModels()
+    {
+        return array_merge(parent::resourceMethodsWithoutModels(), [
+            'get_number_customers_created_this_month',
+            'show_by_user',
+        ]);
+    }
+    */
 }
