+++ api/app/Http/Controllers/UserController.php
@@ -18,10 +18,15 @@ class UserController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('can:viewAny')->only('viewAny');
-        $this->middleware('can:create')->only('create');
-        $this->middleware('can:update')->only('update');
-        $this->middleware('can:delete')->only('delete');
+        //$this->middleware('can:viewAny')->only('viewAny');
+        //$this->middleware('can:create')->only('create');
+        //$this->middleware('can:update')->only('update');
+        //$this->middleware('can:delete')->only('delete');
+        //$this->middleware('can:show_me')->only('show_me');
+        //$this->middleware('can:toogle')->only('toogle');
+        //$this->middleware('can:new_password')->only('new_password');
+
+
 
         /*
         $this->middleware('auth.manager', ['except' => [
@@ -43,6 +48,8 @@ class UserController extends Controller
     }
     public function index(Request $request)
     {
+         /* --- Authorization --- */
+         if (Auth()->guard('api')->user()->type != "EM") { abort(403); }
         $users = $request->type != 'all' ? User::where('type', $request->input('type'))->paginate(10) : User::paginate(10);
         return UserResource::collection($users);
     }
@@ -100,11 +107,15 @@ class UserController extends Controller
 
     public function show(User $user)
     {
+        /* --- Authorization --- */
+        if ((!Auth()->guard('api')->user()->type == "EM") && (Auth()->guard('api')->user()->type !="EM" && Auth()->guard('api')->user()->id!=$user->id) ) { abort(403); }
         return new UserResource($user);
     }
 
     public function update(UpdateUserRequest $request, User $user)
     {
+          /* --- Authorization --- */
+          if ((!Auth()->guard('api')->user()->type == "EM") && (Auth()->guard('api')->user()->type !="EM" && Auth()->guard('api')->user()->id!=$user->id) ) { abort(403); }
         $user->fill($request->validated());
         $user->blocked = $user->blocked==1 ? 0 : 1;
 
@@ -153,6 +164,8 @@ class UserController extends Controller
 
     public function destroy($id) // -> Boolean Return
     {
+         /* --- Authorization --- */
+         if (Auth()->guard('api')->user()->type != "EM") { abort(403); }
         return DB::transaction(function () use ($id) {
             $user = User::where(['id' => $id], ['deleted_at' => null])->firstOrFail();
             if ($user->customer) { $user->customer->delete(); }
@@ -174,10 +187,31 @@ class UserController extends Controller
     }
 
     public function new_password(UpdateUserPasswordRequest $request, User $user) {
+        if (Auth()->guard('api')->user()->type == "EM") { abort(403); }
+
         $user = User::where(['id' => $user->id], ['deleted_at' => null])->firstOrFail();
         $password_hashed = Hash::make($request->input('password'));
         $user->password = $password_hashed;
         $user->save();
         return new UserResource($user);
     }
+/*
+    protected function resourceAbilityMap()
+    {
+        return array_merge(parent::resourceAbilityMap(), [
+            'show_me' => 'show_me',
+            'toogle' => 'toogle',
+            'new_password' => 'new_password'
+    
+        ]);
+    }
+    protected function resourceMethodsWithoutModels()
+    {
+        return array_merge(parent::resourceMethodsWithoutModels(), [
+            'show_me',
+            'toogle',
+            'new_password'
+        ]);
+    }
+    */
 }
